move server to libs
